fix: change GetBalance and GetCode ID from uint32 to int64
Fixed JSON unmarshaling error where the node returns large timestamp-like
IDs (e.g., 1767471866282) that exceed uint32 max value (4,294,967,295).

Error before fix:
  json: cannot unmarshal number 1767471866282 into Go struct field
  GetCode.id of type uint32

Changed affected structs:
- GetBalance.ID: uint32 -> int64
- GetCode.ID: uint32 -> int64

This allows proper parsing of RPC responses from the node when checking
contract code and account balances.

+++ Zond2mongoDB/models/zond.go
@@ -111,13 +111,13 @@ type JsonRPC struct {
 
 type GetBalance struct {
 	JsonRPC string `json:"jsonrpc"`
-	ID      uint32 `json:"id"`
+	ID      int64  `json:"id"`
 	Result  string `json:"result"`
 }
 
 type GetCode struct {
 	JsonRPC string `json:"jsonrpc"`
-	ID      uint32 `json:"id"`
+	ID      int64  `json:"id"`
 	Result  string `json:"result"`
 }
